pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "smart-display"
version = "0.1.0"
description = "Smart display for the Inky Impression 7.3\""
readme = "README.md"
requires-python = ">=3.11"
authors = [{name = "Smart Display Contributors"}]
dependencies = [
    "flask>=3.0",
    "feedparser>=6.0",
    "ics>=0.7",
    "Pillow>=10.3",
    "python-dateutil>=2.9",
    "requests>=2.32"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4"
]
hardware = [
    "inky>=1.5"
]

[tool.setuptools.packages.find]
where = ["src"]
